cast to unsigned

+++ lib/dns/message.c
@@ -46,7 +46,7 @@ hexdump(const char *msg, const char *msg2, void *base, size_t len) {
 	p = base;
 	cnt = 0;
 
-	printf("*** %s [%s] (%u bytes @ %p)\n", msg, msg2, len, base);
+	printf("*** %s [%s] (%u bytes @ %p)\n", msg, msg2, (unsigned)len, base);
 
 	while (cnt < len) {
 		if (cnt % 16 == 0)
